backend/requirements.txt
# Core Django
Django>=5.2,<5.3
django-cors-headers>=4.4,<4.5

# Database & Configuration
psycopg2-binary>=2.9.9,<3.0
python-decouple>=3.8,<4.0

# REST Framework
djangorestframework>=3.15,<3.16
django-filter>=24.3,<25.0
djangorestframework-simplejwt



# Celery
celery>=5.3.6,<5.4
django-celery-results>=2.5,<3.0

# Redis & Caching
redis>=5.0,<5.1
django-redis>=5.4,<5.5

# # HTTP Client
# requests>=2.31,<2.32

# Production WSGI Server
gunicorn>=23.0,<24.0

# Development
# ipython>=8.25,<8.26
django-debug-toolbar>=4.4,<4.5

# Security
django-axes>=6.0,<6.1

# Django dependencies
asgiref>=3.8,<3.9
sqlparse>=0.5,<0.6

